Rearrange parser code
This splits the parser code up into more modules, and renames all the
`Scope` classes to `Parser`. `DeckParser` is now `DeckFilesParser`.
